requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
Werkzeug==2.0.1
